requirements.txt
# 基础 Web 框架依赖
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
gevent>=23.0.0
waitress>=2.1.0

# 配置和数据处理
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
psutil>=5.9.0
requests>=2.31.0

# OpenAI 兼容性
openai>=1.0.0
tiktoken>=0.7.0

# 数据处理和科学计算
numpy>=1.24.0
torch>=2.1.0
transformers>=4.35.0
python-dotenv

# 跨平台推理引擎 (条件安装)
vllm>=0.2.0; sys_platform != "darwin"
mlx-lm>=0.5.0; sys_platform == "darwin" and platform_machine == "arm64"
llama-cpp-python>=0.2.0

# 模型下载和管理
modelscope>=1.9.0
huggingface-hub>=0.17.0
safetensors>=0.4.0

# 测试和开发工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

